% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{beanPlotFeatures}
\alias{beanPlotFeatures}
\title{beanPlotFeatures}
\usage{
beanPlotFeatures(data, target, subsetvariable = NULL, col = c("orange",
  "blue"), quantileCut = F, interquantile = 10, ...)
}
\description{
plots a beanplot ot two classes
}
