% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineModel.R
\name{combineModel}
\alias{combineModel}
\title{combineModel}
\usage{
combineModel(trainOutput, featureFunctionList, test.data = NULL,
  threshold = NULL, positveClass = NULL)
}
\arguments{
\item{trainOutput:}{a list containing a model returned by generateModel.}

\item{featureFunctionList:}{a list of functions returned by generateFunctionList.}

\item{test.data:}{a data.frame containing the data used for testing the model.}

\item{threshold:}{an optional single numeric value that represents the threshold that should be applied during the prediction.}

\item{positveClass:}{a character vector of length one representing a string which specifies a positive example}
}
\description{
This function combines the output from generateModel with other parameters to a model of class WrappedCombiModel.
}
