% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluateFunList}
\alias{evaluateFunList}
\title{evaluateFunList}
\usage{
evaluateFunList(funList, data)
}
\arguments{
\item{funlist:}{a list that contains functions from whom the first argument is data.}

\item{data:}{a data.frame on which the functions from funList should be evaluated on.}
}
\description{
Returns a data.frame that consist of the applied data and the additional features calculated with the functions from funlist.
}
