% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModel.R
\name{generateModel}
\alias{generateModel}
\title{generateModel}
\usage{
generateModel(classifier, us.rate, features, data, targetVariable,
  positiveClass, hyperpars = list(), estimatingThreshold = F,
  tprThreshold = 0.995)
}
\arguments{
\item{classifier:}{a character vector of length one specifying the classification algorithm to use. The name has to be applied as specified in the mlr package e.g. "classif.randomForest"}

\item{us.rate:}{a single numeric value between 0 and 1 that specifies the undersampling rate that is applied}

\item{features:}{a character vector representing column names that are used for model training}

\item{data:}{a data.frame, which should be used for training. If additional features were calculated data have to contain them already.}

\item{targetVariable:}{a character vector of length one representing the column that should be used as target.}

\item{positiveClass:}{a character vector of length one representing a string which specifies a positive example}

\item{hyperpars:}{a list of hyper parameters passed to the mlr model}

\item{estimatingThreshold:}{a boolean vector of length one specifying whether the threshold should be estimated or not.}

\item{tprThreshold:}{a single numeric value between 0 and 1 that specifies the true positive rate onto which the threshold should be tuned.}
}
\description{
This function returns a model a list of with the trained model and their parameters.
}
