% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getThresholdCV}
\alias{getThresholdCV}
\title{getThresholdCV}
\usage{
getThresholdCV(resampleResult, train.data, tprThreshold = 0.99)
}
\arguments{
\item{resampleResult:}{The result of the resample function from the mlr package.}

\item{train.data:}{The data.frame used to train the model.}

\item{tprThreshold:}{A single numeric value specifying towards which true positive rate the threshold should be tuned.}
}
\description{
This function returns a single numeric value, that represents the optimal threshold.
}
