% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.WrappedCombiModel}
\alias{predict.WrappedCombiModel}
\title{predict.WrappedCombiModel}
\usage{
\method{predict}{WrappedCombiModel}(combinedModel, newdata, NAtoZero = F, ...)
}
\arguments{
\item{combinedModel:}{a model of class WrappedCombiModel.}

\item{newdata:}{a data.frame containing new data that should be predicted.}

\item{NAtoZero:}{if TRUE, all observations that contain a NA in one of the features used by the model will be predicted with a probability of 0. If FALSE, for those observations NA is returned.}
}
\description{
This function returns a prediction object from the package mlr.
}
