% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrain.R
\name{retrain}
\alias{retrain}
\title{retrain}
\usage{
retrain(combinedModel, newdata, estimatingThreshold = F,
  tprThreshold = 0.995, keepData = F)
}
\arguments{
\item{combinedModel:}{a model of class WrappedCombiModel.}

\item{newdata:}{a data.frame containing new data that should be used for retraining.}

\item{estimatingThreshold:}{a boolean vector of length one specifying whether the threshold should be estimated or not.}

\item{tprThreshold:}{a single numeric value between 0 and 1 that specifies the true positive  rate onto which the threshold should be tuned.}

\item{keepData:}{if TRUE the data stored in the combinedModel were kept. In this case the layout of the data has to be identical. If FALSE only newdata is stored in the retrained model and the retraining is based only on this data.}
}
\description{
This function retrains a existing model of class WrappedCombiModel and returns it.
}
